Refactored repositories into a new package
